requirements.txt
absl-py==2.0.0
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.0.1
click==8.1.3
docopt==0.6.2
Flask==2.2.3
flatbuffers==23.5.26
gast==0.4.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.2
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.1
pipreqs==0.4.11
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
threadpoolctl==3.2.0
typing-extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
zipp==3.15.0
